+++ src/komponenter/header/meny/ekspanderende-menyer/varsler-dropdown-desktop/varselvisning/Varsel.less
@@ -1,0 +1,28 @@
+
+.varsel {
+    &__container {
+        display: flex;
+        flex-direction: column;
+    }
+
+    &__ikon-container {
+        //
+    }
+
+    &__innhold {
+        display: flex;
+        flex-direction: row;
+    }
+
+    &__datotid {
+        //
+    }
+
+    &__melding {
+        //
+    }
+
+    &__lenke {
+        //
+    }
+}
