+++ src/komponenter/header/meny/varsel/varselvisning/Varselvisning.tsx
@@ -1,0 +1,180 @@
+import React from 'react';
+import { connect } from 'react-redux';
+import parse from 'html-react-parser';
+import { AppState } from '../../../../../reducer/reducer';
+import { desktopview, tabletview } from '../../../../../styling-mediaquery';
+import Environment from '../../../../../utils/Environment';
+import './VarselvisningOld.less';
+import {
+    GACategory,
+    triggerGaEvent,
+} from '../../../../../utils/google-analytics';
+import Tekst, { finnTekst } from '../../../../../tekster/finn-tekst';
+import { Language } from '../../../../../reducer/language-duck';
+import { LenkeMedGA } from '../../../../LenkeMedGA';
+
+interface OwnProps {
+    tabIndex: boolean;
+    togglevarselmeny?: () => void;
+}
+
+interface StateProps {
+    varsler: string;
+    antallVarsler: number;
+    antallUlesteVarsler: number;
+    language: Language;
+}
+
+interface State {
+    parsedVarsler: any;
+    windowSize: number;
+}
+
+type Props = OwnProps & StateProps;
+
+class Varselvisning extends React.Component<Props, State> {
+    constructor(props: Props) {
+        super(props);
+
+        this.state = {
+            parsedVarsler: parse(this.props.varsler),
+            windowSize: window.innerWidth,
+        };
+        this.handleWindowSize = this.handleWindowSize.bind(this);
+    }
+
+    handleWindowSize() {
+        this.setState({
+            windowSize: window.innerWidth,
+        });
+    }
+
+    setTabIndex = () => {
+        const varslerWrapperElement: string = this.erTabletEllerDesktop()
+            ? '.desktopmeny .nav-varsler'
+            : '.mobilmeny .nav-varsler';
+
+        const varsler = document.querySelector(varslerWrapperElement);
+
+        if (varsler) {
+            for (
+                let i = 0;
+                i <= varsler.getElementsByTagName('a').length;
+                i++
+            ) {
+                if (varsler.getElementsByTagName('a')[i]) {
+                    varsler.getElementsByTagName('a')[i].tabIndex = this.props
+                        .tabIndex
+                        ? 0
+                        : -1;
+                }
+            }
+        }
+    };
+
+    gaEventHandler = (event: Event) => {
+        if (event.type === 'auxclick' && (event as MouseEvent).button !== 1) {
+            return;
+        }
+        triggerGaEvent({
+            category: GACategory.Header,
+            action: 'varsel-lenke',
+            label: (event.target as HTMLAnchorElement).href,
+        });
+    };
+
+    getVarselLenker = () => document.getElementsByClassName('varsel-lenke');
+
+    addGaEventTriggers = () => {
+        for (const varselLenke of this.getVarselLenker()) {
+            if (varselLenke.tagName.toUpperCase() === 'A') {
+                varselLenke.addEventListener('click', this.gaEventHandler);
+                varselLenke.addEventListener('auxclick', this.gaEventHandler);
+            }
+        }
+    };
+
+    removeGaEventTriggers = () => {
+        for (const varselLenke of this.getVarselLenker()) {
+            if (varselLenke.tagName.toUpperCase() === 'A') {
+                varselLenke.removeEventListener('click', this.gaEventHandler);
+                varselLenke.removeEventListener(
+                    'auxclick',
+                    this.gaEventHandler
+                );
+            }
+        }
+    };
+
+    componentDidMount(): void {
+        window.addEventListener('resize', this.handleWindowSize);
+        this.addGaEventTriggers();
+        this.setTabIndex();
+    }
+
+    componentDidUpdate(prevProps: Readonly<Props>): void {
+        if (this.props.togglevarselmeny) {
+            this.setTabIndex();
+            if (this.props.tabIndex && !prevProps.tabIndex) {
+                this.props.togglevarselmeny();
+            }
+        }
+    }
+
+    componentWillUnmount() {
+        window.removeEventListener('resize', this.handleWindowSize);
+        this.removeGaEventTriggers();
+    }
+
+    erTabletEllerDesktop = () => {
+        return this.state.windowSize > tabletview - 1;
+    };
+
+    erDesktop = () => {
+        return this.state.windowSize > desktopview - 1;
+    };
+
+    render() {
+        const { tabIndex, antallUlesteVarsler, antallVarsler } = this.props;
+        const klassenavn = this.erDesktop()
+            ? 'varsler-display-desktop'
+            : 'varsler-display-mobil-tablet';
+
+        return (
+            <div className={klassenavn}>
+                {this.state.parsedVarsler}
+
+                {antallVarsler > 5 && (
+                    <div className="vis-alle-lenke skillelinje-topp">
+                        <LenkeMedGA
+                            href={Environment.API_VARSELINNBOKS_URL}
+                            tabIndex={tabIndex ? 0 : -1}
+                            gaEventArgs={{
+                                category: GACategory.Header,
+                                action: 'varsler/visalle',
+                                label: Environment.API_VARSELINNBOKS_URL,
+                            }}
+                        >
+                            <Tekst id={'varsler-visalle'} />
+                            {antallUlesteVarsler > 0
+                                ? ` (${antallUlesteVarsler} ${finnTekst(
+                                    'varsler-nye',
+                                    this.props.language
+                                )})`
+                                : ''}
+                        </LenkeMedGA>
+                    </div>
+                )}
+            </div>
+        );
+    }
+}
+
+const mapStateToProps = (state: AppState): StateProps => ({
+    varsler: state.varsler.data.varsler,
+    antallVarsler: state.varsler.data.antall,
+    antallUlesteVarsler: state.varsler.data.uleste,
+    language: state.language.language,
+});
+
+export default connect(mapStateToProps)(Varselvisning);
