+++ src/komponenter/header/meny/varsel/varselvisning/VarselvisningMobil.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import BEMHelper from '../../../../../utils/bem';
 import Lukkundermeny from '../../ekspanderende-menyer/hovedmeny-mobil/meny-dropdown/mobil-innhold/Lukkundermeny';
-import VarselVisning from '../../ekspanderende-menyer/varsler-dropdown-desktop/varselvisning/Varselvisning';
+import VarselVisning from './Varselvisning';
 import TopSeksjon from '../../ekspanderende-menyer/hovedmeny-mobil/meny-dropdown/mobil-innhold/top-seksjon/Topseksjon';
-import './VarselvisningMobil.less';
+import './VarselvisningOld.less';
 
 interface OwnProps {
     visvarsel: boolean;
